+++ PSFramework/internal/scripts/strings.ps1
@@ -1,4 +1,7 @@
-﻿Import-LocalizedString -Path (Resolve-Path "$script:ModuleRoot\en-us\stringsAssembly.psd1") -Module PSFramework -Language 'en-US'
-
+﻿# Load English Language Files
+foreach ($item in (Resolve-Path "$script:ModuleRoot\en-us\*.psd1"))
+{
+	Import-LocalizedString -Path $item -Module PSFramework -Language 'en-US'
+}
 
 $script:strings = Get-PSFLocalizedString -Module PSFramework